Moved story + smoke
